feat(python-cli): separated rpc and cli in python
Separated RPC-related code and CLI-related code into distinct packages
