+++ perun-cli-python/perun/cli/getUser.py
@@ -1,6 +1,6 @@
 from perun_openapi import ApiException
 from rich import print
-from perun.cli import PerunException
+from perun.rpc import PerunException
 import typer
 import perun.cli
